compat: gen: add support for 'absent or lacks' combination
Currently we differentiate between something being 'absent', like when
there is no definition of given struct, and when something 'lacks' given
feature, like when given struct is present, but has no field we are
looking for.
Add suport for the case when we don't care if given thing is entriely
missing or incompatible for us, like in the case of missing macro
invocation.

main awk program that actually does the pattern matching have very similar
logic for 'absent' and 'lacks' operators, the difference is that the
latter requires non-empty input (IOW that we have found the thing we match
against). With the new 'absent or lacks' combination of operators
(combination as seen from gen DSL) we pass the operator as 'absent', what
means that we just remove the "!empty" part of the logic; but with an
actual pattern (passed to "lacks").

Change-type: FeatureImplementation
Signed-off-by: Przemek Kitszel <przemyslaw.kitszel@intel.com>