idpf-linux: fix IDPF_ADD_PROBES idpf_tx_extra_counters NULL deref
The recent Tx buffer pool refactor moved the initialization of the first
buffer lower in splitq_frame. However, the call to
idpf_tx_extra_counters derefs the first buffer for some of the
additional counters. It also uses the TSO related fields in the offload
struct for some counters as well which would not have been initializaed
either.

Move the extra_counters call to almost the very end of splitq_frame to
ensure everything is initialized properly before modifying any counters.

Signed-off-by: Joshua Hay <joshua.a.hay@intel.com>
Change-type: DefectResolution